moved generic name
