+++ scanomatic/util/generic_name.py
